refactor: apiservice and useHandleError were passed to the shared folder
