+++ cdd/src/app/(footered)/enviar-sacola/components/SendBagTable.tsx
@@ -7,11 +7,10 @@ import { ChangeEvent, useEffect, useState } from "react";
 import { toast } from "sonner";
 import { HiOutlineSearch } from "react-icons/hi";
 import Button from "@shared/components/Button";
-import { useGetLocalStorage } from "@cdd/app/hooks/useGetLocalStorage";
-import RedirectModal from "@cdd/components/SessionExpiredModal";
-import { Bag } from "@cdd/interfaces/bag";
+import { useLocalStorage } from "@shared/hooks/useLocalStorage"
+import { Bag } from "@shared/interfaces/bag"
 import Loader from "@shared/components/Loader";
-import { useHandleError } from "@cdd/app/hooks/useHandleError";
+import { useHandleError } from "@shared/hooks/useHandleError";
 
 interface BagsProps {
   page: number;
@@ -25,6 +24,7 @@ export default function SendBagTable({ page }: BagsProps) {
   const [isLoading, setIsLoading] = useState(false);
 
   const { handleError } = useHandleError()
+  const { getFromStorage } = useLocalStorage()
 
   const handleChangeSearchInput = (e: ChangeEvent<HTMLInputElement>) => {
     setTimeout(() => {
@@ -35,7 +35,7 @@ export default function SendBagTable({ page }: BagsProps) {
   useEffect(() => {
     (async () => {
       setIsLoading(true);
-      const cycle = useGetLocalStorage("selected-cycle");
+      const cycle = getFromStorage("selected-cycle");
 
       if (!cycle) {
         toast.error("Selecione um ciclo para ver os pedidos!");
