+++ cdd/src/app/_actions/bag/fetch-bag.ts
@@ -1,6 +1,6 @@
 "use server"
 
-import ApiService from "@cdd/service/index"
+import ApiService from "@shared/service/index"
 
 interface FetchBagRequest {
   bag_id: string
