+++ cdd/src/app/_actions/bag/list-bags.ts
@@ -1,6 +1,6 @@
 "use server"
 
-import ApiService from "@cdd/service/index"
+import ApiService from "@shared/service/index"
 
 interface ListBagsRequest {
   page: number;
