+++ cdd/src/app/_actions/order/get-order.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import ApiService from "@cdd/service/index"
+import ApiService from "@shared/service/index"
 
 export async function getOrder(order_id: string) {
   const response = ApiService.GET({
