--- cdd/src/app/hooks/useSetLocalStorage.ts
@@ -1,5 +1,0 @@
-export function useSetLocalStorage(key: string, value: any){
-  const valueString = JSON.stringify(value)
-
-  localStorage.setItem(key, valueString)
-}
