+++ shared/src/errors.ts
