+++ shared/src/interfaces/bag.ts
